Filter sort pagination feature

+++ Controllers/CommentController.cs
@@ -19,37 +19,41 @@ public class CommentController : ControllerBase {
     
     [HttpGet]
     public async Task<IActionResult> GetAllComment() {
-        var listComment = await _commentRepository.GetAllComment();
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        var listComment = await _commentRepository.GetAllCommentAsync();
         var commentDto = listComment.Select (c => c.toCommentDTO());
         return Ok(commentDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetCommentById([FromRoute] int id) {
         var listComment = await _commentRepository.GetCommentByIdAsync(id);
         return listComment == null ? NotFound() : Ok(listComment);
     }
 
-    [HttpPost("{stockId}")]
+    [HttpPost("{stockId:int}")]
     public async Task<IActionResult> CreteNewComment([FromRoute] int stockId,[FromBody] CreateCommentDTO commentDto) {
         if (!await _stockRepository.isStockExists(stockId)) {
             return BadRequest("Stock does not exist.");
         }
 
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var commentModel = commentDto.toCommentFromCreateDTO(stockId);
         await _commentRepository.CreateNewCommentAsync(commentModel);
         return CreatedAtAction(nameof(GetCommentById), new {id = commentModel.Id}, commentModel.toCommentDTO());
     }
 
-    [HttpPut("{commentId}")]
+    [HttpPut("{commentId:int}")]
     public async Task<IActionResult> UpdateComment([FromRoute] int commentId, [FromBody] UpdateCommentDTO commentDto) {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var commentModel = commentDto.toCommentFromUpdateDTO();
         var result = await _commentRepository.UpdateCommentAsync(commentId, commentModel);
         return result == null ? NotFound() : Ok(result);
     }
 
-    [HttpDelete("{commentId}")]
+    [HttpDelete("{commentId:int}")]
     public async Task<IActionResult> DeleteComment([FromRoute] int commentId) {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
         var result = await _commentRepository.DeleteCommentAsync(commentId);
         return result == null ? NotFound() : NoContent();
